pyproject.toml
[tool.poetry]
name = "university-website-scraper"
version = "0.1.0"
description = ""
authors = ["Dylan Medina <dylan@dylanmedina.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "university_website_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
MechanicalSoup = "^1.3.0"
beautifulsoup4 = "^4.12.3"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
